refactor(style): 공통styled-component layout으로 분리
